requirements.txt
pyhtcc>=0.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
flask>=3.0.0
apscheduler>=3.10.0
gunicorn>=21.0.0
requests>=2.31.0
